[*] Moving Ai Tests in the root tests directory
